+++ src/buildURIFragment.ts
@@ -1,6 +1,13 @@
 import { DelegationChain } from '@dfinity/identity';
 import { buildDelegationString } from './buildDelegationString';
 
+type BuildURIFragmentParams = {
+  /** The delegation chain to encode */
+  delegationChain: DelegationChain;
+  /** The session ID to include in the URI fragment */
+  sessionId: string;
+};
+
 /**
  * Builds a URI fragment containing the encoded delegation information.
  * This is used for secure transmission of delegation data in the URL fragment
@@ -9,8 +16,11 @@ import { buildDelegationString } from './buildDelegationString';
  * @param {DelegationChain} delegationChain - The delegation chain containing the delegation information.
  * @returns {string} A URI fragment in the format 'delegation=<encoded_delegation_string>'.
  */
-export const buildURIFragment = (delegationChain: DelegationChain): string => {
+export const buildURIFragment = ({
+  delegationChain,
+  sessionId,
+}: BuildURIFragmentParams): string => {
   const delegationString = buildDelegationString(delegationChain);
   const encodedDelegation = encodeURIComponent(delegationString);
-  return `delegation=${encodedDelegation}`;
+  return `delegation=${encodedDelegation}&session-id=${sessionId}`;
 };